Refactor: move DhisService tests back into single file
